formatted backend structure and created addUserBundle Endpoint
